+++ Sources/KeyboardKit/Appearance/KeyboardBackgroundStyle.swift
@@ -11,9 +11,10 @@ import SwiftUI
 /**
  This style defines the background of a keyboard.
 
- Although the style only has a ``backgroundType``, these two
- types should be separated, since the style will be extended
- with more properties in the future, while the type will not.
+ This style has many components that can be used to create a
+ background view that can then be added as a background to a
+ keyboard view. The ``backgroundView`` property returns such
+ a view, which applies all properties in a `ZStack`.
 
  You can modify the ``standard`` style to change the default,
  global background style.
@@ -21,21 +22,55 @@ import SwiftUI
 public struct KeyboardBackgroundStyle: Codable, Equatable {
 
     /**
-     Create a keyboard background style.
+     Create a background style with optional components.
 
      - Parameters:
-        - type: The background type to use, by default `.clear`.
+       - backgroundColor: A background color to apply, by default `nil`.
+       - backgroundGradient: A background gradient color set, by default `nil`.
+       - imageData: Optional image data that can be used to create a background image, by default `nil`.
+       - overlayColor: An overlay color to apply, by default `nil`.
+       - overlayGradient: An overlay gradient color set, by default `nil`.
      */
     public init(
-        _ type: KeyboardBackgroundType = .clear
+        backgroundColor: Color? = nil,
+        backgroundGradient: [Color]? = nil,
+        imageData: Data? = nil,
+        overlayColor: Color? = nil,
+        overlayGradient: [Color]? = nil
     ) {
-        self.backgroundType = type
+        self.backgroundColor = backgroundColor
+        self.backgroundGradient = backgroundGradient
+        self.imageData = imageData
+        self.overlayColor = overlayColor
+        self.overlayGradient = overlayGradient
     }
 
-    /**
-     The background type to use.
-     */
-    public var backgroundType: KeyboardBackgroundType
+    @available(*, deprecated, message: "Use the component initializer instead.")
+    public init(_ type: KeyboardBackgroundType) {
+        self.backgroundType = type
+        self.legacyBackgroundType = type
+    }
+
+    private var legacyBackgroundType: KeyboardBackgroundType = .none
+
+    @available(*, deprecated, message: "Use the component properties instead.")
+    public var backgroundType: KeyboardBackgroundType = .none
+
+
+    /// A background color to apply.
+    public var backgroundColor: Color?
+
+    /// A background gradient color set.
+    public var backgroundGradient: [Color]?
+
+    /// Optional image data that can be used to create a background image.
+    public var imageData: Data?
+
+    /// An overlay color to apply.
+    public var overlayColor: Color?
+
+    /// An overlay gradient color set.
+    public var overlayGradient: [Color]?
 }
 
 public extension KeyboardBackgroundStyle {
@@ -48,10 +83,40 @@ public extension KeyboardBackgroundStyle {
     static var standard = KeyboardBackgroundStyle()
 
     /**
-     The background view to use for the style.
+     Create a background view that uses all style properties.
      */
     @ViewBuilder
     var backgroundView: some View {
-        backgroundType.view
+        if legacyBackgroundType != .none {
+            legacyBackgroundType.internalView
+        } else {
+            ZStack {
+                backgroundColor
+                if let backgroundGradient {
+                    LinearGradient(colors: backgroundGradient, startPoint: .top, endPoint: .bottom)
+                }
+                image(from: imageData)
+                if let overlayGradient {
+                    LinearGradient(colors: overlayGradient, startPoint: .top, endPoint: .bottom)
+                }
+                overlayColor
+            }
+        }
+    }
+}
+
+private extension KeyboardBackgroundStyle {
+
+    func image(from data: Data?) -> Image? {
+        guard let data else { return nil }
+    #if canImport(UIKit)
+        let image: UIImage = UIImage(data: data) ?? UIImage()
+        return Image(uiImage: image)
+    #elseif canImport(AppKit)
+        let image: NSImage = NSImage(data: data) ?? NSImage()
+        return Image(nsImage: image)
+    #else
+        return Image(systemImage: "exclamationmark.triangle")
+    #endif
     }
 }